update addHits function to native upsert and update query
