+++ _tracking/tracking/src/main/java/com/example/tracking/Controller/TrackingController.java
@@ -2,7 +2,6 @@ package com.example.tracking.Controller;
 
 import com.example.tracking.DTO.HistoryDTO;
 import com.example.tracking.DTO.HitsDTO;
-import com.example.tracking.Service.DailyLockFacade;
 import com.example.tracking.Service.TrackingService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -20,13 +19,12 @@ import java.util.Map;
 @RestController @RequiredArgsConstructor
 public class TrackingController {
     private final TrackingService trackingService;
-    private final DailyLockFacade dailyLockFacade;
 
     @Operation(summary = "url 조회수 증가", description = "해당 url이 존재하면 일일 조회수와 총 조회수를 증가하고, 아니면 url 조회수 정보를 생성합니다.")
     @PutMapping("/hits")
     public ResponseEntity<Object> addHits(@Parameter(name = "url", description = "url을 작성합니다.(':'는 '%3A', '/'는 '%2F'로 작성)")
                                               @RequestParam String url){
-        this.dailyLockFacade.addHits(url);
+        this.trackingService.addHits(url);
 
         Map<String, String> response = new HashMap<>(){};
         response.put("result", "success");
