+++ _tracking/tracking/src/test/java/com/example/tracking/TrackingServiceTest.java
@@ -145,4 +145,22 @@ public class TrackingServiceTest {
         long end = System.currentTimeMillis();
         System.out.println((double)(end-start)/1000 + "seconds");
     }
+
+    @Test
+    public void addHitsTestWithLargeTraffic() throws InterruptedException {
+        //thread 생성
+        ExecutorService executorService = Executors.newFixedThreadPool(10);
+        CountDownLatch countDownLatch = new CountDownLatch(10);
+
+        //조회수 증가 테스트
+        this.dailyRepository.save(new Daily("www.google.com", 1, 1L));
+        for(int i = 0; i < 10; i++){
+            executorService.execute(()->{
+                this.trackingService.addHits("www.google.com");
+                countDownLatch.countDown();
+            });
+        }
+        countDownLatch.await();
+        assertEquals(11, this.dailyRepository.findByUrl("www.google.com").get().getTodayHit());
+    }
 }
